+++ influxdb_client/client/_base.py
@@ -60,6 +60,10 @@ class _BaseClient(object):
         self.conf.enable_gzip = enable_gzip
         self.conf.verify_ssl = kwargs.get('verify_ssl', True)
         self.conf.ssl_ca_cert = kwargs.get('ssl_ca_cert', None)
+        self.conf.cert_file = kwargs.get('cert_file', None)
+        self.conf.key_file = kwargs.get('key_file', None)
+        self.conf.key_password = kwargs.get('key_password', None)
+        self.conf.ssl_context = kwargs.get('ssl_context', None)
         self.conf.proxy = kwargs.get('proxy', None)
         self.conf.proxy_headers = kwargs.get('proxy_headers', None)
         self.conf.connection_pool_maxsize = kwargs.get('connection_pool_maxsize', self.conf.connection_pool_maxsize)
@@ -142,6 +146,18 @@ class _BaseClient(object):
         if _has_option('ssl_ca_cert'):
             ssl_ca_cert = _config_value('ssl_ca_cert')
 
+        cert_file = None
+        if _has_option('cert_file'):
+            cert_file = _config_value('cert_file')
+
+        key_file = None
+        if _has_option('key_file'):
+            key_file = _config_value('key_file')
+
+        key_password = None
+        if _has_option('key_password'):
+            key_password = _config_value('key_password')
+
         connection_pool_maxsize = None
         if _has_option('connection_pool_maxsize'):
             connection_pool_maxsize = _config_value('connection_pool_maxsize')
@@ -168,6 +184,7 @@ class _BaseClient(object):
 
         return cls(url, token, debug=debug, timeout=_to_int(timeout), org=org, default_tags=default_tags,
                    enable_gzip=enable_gzip, verify_ssl=_to_bool(verify_ssl), ssl_ca_cert=ssl_ca_cert,
+                   cert_file=cert_file, key_file=key_file, key_password=key_password,
                    connection_pool_maxsize=_to_int(connection_pool_maxsize), auth_basic=_to_bool(auth_basic),
                    profilers=profilers, proxy=proxy, **kwargs)
 
@@ -179,6 +196,9 @@ class _BaseClient(object):
         org = os.getenv('INFLUXDB_V2_ORG', "my-org")
         verify_ssl = os.getenv('INFLUXDB_V2_VERIFY_SSL', "True")
         ssl_ca_cert = os.getenv('INFLUXDB_V2_SSL_CA_CERT', None)
+        cert_file = os.getenv('INFLUXDB_V2_CERT_FILE', None)
+        key_file = os.getenv('INFLUXDB_V2_KEY_FILE', None)
+        key_password = os.getenv('INFLUXDB_V2_KEY_PASSWORD', None)
         connection_pool_maxsize = os.getenv('INFLUXDB_V2_CONNECTION_POOL_MAXSIZE', None)
         auth_basic = os.getenv('INFLUXDB_V2_AUTH_BASIC', "False")
 
@@ -195,6 +215,7 @@ class _BaseClient(object):
 
         return cls(url, token, debug=debug, timeout=_to_int(timeout), org=org, default_tags=default_tags,
                    enable_gzip=enable_gzip, verify_ssl=_to_bool(verify_ssl), ssl_ca_cert=ssl_ca_cert,
+                   cert_file=cert_file, key_file=key_file, key_password=key_password,
                    connection_pool_maxsize=_to_int(connection_pool_maxsize), auth_basic=_to_bool(auth_basic),
                    profilers=profilers, **kwargs)