+++ influxdb_client/configuration.py
@@ -90,8 +90,13 @@ class Configuration(object, metaclass=TypeWithDefault):
         self.cert_file = None
         # client key file
         self.key_file = None
+        # client key file password
+        self.key_password = None
         # Set this to True/False to enable/disable SSL hostname verification.
         self.assert_hostname = None
+
+        # Set this to specify a custom ssl context to inject this context inside the urllib3 connection pool.
+        self.ssl_context = None
 
         # urllib3 connection pool's maximum number of connections saved
         # per pool. urllib3 uses 1 connection as default value, but this is